refactor(reporter): move shared wire leaf shapes + method list to core
Extract the five identical wire leaf shapes (SuiteConfigEntry, TestAnnotation,
FilterDetails, BrowserConfig, TestStepEvent) and the canonical locator-builder
method list into @piwitests/core. The reporter's wire.ts and the app's types.ts
now re-export the leaf shapes from core; error-fingerprint's LEAF_SELECTOR_METHODS
and the reporter's LOCATOR_METHODS derive from the shared list. The per-case
payload shapes (WireTestCase vs TestCasePayload/StreamEventPayload) stay separate
and keep their wire drift-guard test, since they are intentionally different.
